+++ javalib/src/main/scala/java/lang/Double.scala
@@ -248,7 +248,8 @@ object Double {
       errno.errno = 0
       val res = stdlib.strtod(cstr, end)
 
-      if (errno.errno == 0 && cstr != !end && string.strlen(!end) == 0) res
+      if (errno.errno == 0 && cstr != !end && string.strlen(!end).toInt == 0)
+        res
       else throw new NumberFormatException(s)
     }
